refactor: split extension module responsibilities
